feat: Improve rolling_(sum|mean) accuracy
