+++ crates/polars-compute/src/rolling/no_nulls/mean.rs
@@ -1,41 +1,7 @@
 #![allow(unsafe_op_in_unsafe_fn)]
 
+use super::super::mean::MeanWindow;
 use super::*;
-
-pub struct MeanWindow<'a, T> {
-    sum: SumWindow<'a, T, f64>,
-}
-
-impl<'a, T> RollingAggWindowNoNulls<'a, T> for MeanWindow<'a, T>
-where
-    T: NativeType
-        + IsFloat
-        + std::iter::Sum
-        + AddAssign
-        + SubAssign
-        + Div<Output = T>
-        + NumCast
-        + Add<Output = T>
-        + Sub<Output = T>
-        + PartialOrd,
-{
-    fn new(
-        slice: &'a [T],
-        start: usize,
-        end: usize,
-        params: Option<RollingFnParams>,
-        window_size: Option<usize>,
-    ) -> Self {
-        Self {
-            sum: SumWindow::<T, f64>::new(slice, start, end, params, window_size),
-        }
-    }
-
-    unsafe fn update(&mut self, start: usize, end: usize) -> Option<T> {
-        let sum = self.sum.update(start, end).unwrap_unchecked();
-        Some(sum / NumCast::from(end - start).unwrap())
-    }
-}
 
 pub fn rolling_mean<T>(
     values: &[T],
